requirements.txt
tkinter
mysql-connector-python
pandas
matplotlib
Pillow
reportlab